requirements.txt
numpy>=1.13.3
torch==1.7.0
transformers==4.11.3
sklearn
matplotlib
blitz-bayesian-pytorch